requirements.txt
pandas>=2.0.0
numpy>=1.24.0
django>=5.0.0
plotly>=5.17.0
openai>=1.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
